+++ index.js
@@ -1,242 +1,297 @@
 #!/usr/bin/env node
 
-const childProcess = require('child_process');
-const request = require('request');
-const path = require('path');
-const rimraf = require('rimraf');
-const fs = require('fs');
-const showdown  = require('showdown');
+const childProcess = require("child_process");
+const request = require("request");
+const path = require("path");
+const rimraf = require("rimraf");
+const fs = require("fs");
+const showdown = require("showdown");
 
-const staticDirectory = path.join(__dirname, 'static');
+const staticDirectory = path.join(__dirname, "static");
 const lockFile = readLockFile();
-const converter = new showdown.Converter();
+const outputDirectory = getOutputDirectoryPath();
 
-let outputDirectory = './reports/outdated-changelogs';
-let changelogs = [];
+const converter = new showdown.Converter();
+converter.setFlavor("github");
 
 main();
-
 
 // main
 // ===================
 
 async function main() {
-  parseParameters();
-  setupMarkdoneConverter();
-  const dependencies = await getOutdatedDependencies();
+  const allDependencyList = await getOutdatedDependencies();
+  const allKeyList = Object.keys(allDependencyList);
   clearOutputDirectory();
-  checkNextDependency(dependencies);
-}
 
-function parseParameters() {
-  const outputParamIndex = process.argv.findIndex((value) => value === '-o')
-  if (outputParamIndex !== -1) {
-    const outputParamValue = process.argv.slice(outputParamIndex+1, outputParamIndex+2)[0];
-    outputDirectory = outputParamValue || outputDirectory;
-    console.log(outputDirectory)
+  let i = 0;
+  const allChangelogList = [];
+  for (const key of allKeyList) {
+    i++;
+    const dependency = allDependencyList[key];
+
+    console.log(``);
+    logMessage(`${key} (${i}/${allKeyList.length})`);
+
+    const result = await checkDependency(i, key, dependency);
+    allChangelogList.push(result);
+
+    await sleep(1000);
   }
+
+  writeFiles(allChangelogList);
 }
 
-function setupMarkdoneConverter() {
-  converter.setFlavor('github');
-}
+function getOutputDirectoryPath() {
+  let result = "./reports/outdated-changelogs";
 
+  const outputParamIndex = process.argv.findIndex((value) => value === "-o");
+  if (outputParamIndex !== -1) {
+    const outputParamValue = process.argv.slice(
+      outputParamIndex + 1,
+      outputParamIndex + 2
+    )[0];
+
+    result = outputParamValue || result;
+  }
+
+  return result;
+}
 
 // npm
 // ===================
 
 function readLockFile() {
-  logMessage('read lock file');
-  let rawdata = fs.readFileSync('package-lock.json');
+  logMessage("read lock file");
+  const rawdata = fs.readFileSync("package-lock.json");
   return JSON.parse(rawdata);
 }
 
-function getOutdatedDependencies () {
-  logMessage('get outdated dependencies');
-	return new Promise((resolve, reject) => {
-		childProcess.exec('npm outdated --json --long', (error, stdout) => {
-			if (error && stdout.length === 0) {
-				reject(error);
-      }
-      
-      else {
-        let response = JSON.parse(stdout || '{}')
+async function getOutdatedDependencies() {
+  logMessage("get outdated dependencies");
+  return new Promise((resolve, reject) => {
+    childProcess.exec("npm outdated --json --long", (error, stdout) => {
+      if (error && stdout.length === 0) {
+        reject(error);
+      } else {
+        const response = JSON.parse(stdout || "{}");
         resolve(response);
       }
-		});
-	});
+    });
+  });
 }
 
-function getRegistryInfos(homepageUrl, dependencyName, currentVersion, targetVersion) {
+async function getRegistryInfos(
+  homepageUrl,
+  dependencyName,
+  currentVersion,
+  targetVersion
+) {
   logInfo(`get infos for ${dependencyName}`);
 
-	return new Promise((resolve, reject) => {
-    let registryUrl = getRegistryUrl(dependencyName);
-    request(registryUrl, {json: true}, (err, res, json) => {
-      if (res && res.statusCode < 300 && json && json['repository'] && json['repository']['url']) {
+  return new Promise((resolve, reject) => {
+    const registryUrl = getRegistryUrl(dependencyName);
+    request(registryUrl, { json: true }, (err, res, json) => {
+      if (
+        res &&
+        res.statusCode < 300 &&
+        json &&
+        json["repository"] &&
+        json["repository"]["url"]
+      ) {
         resolve({
-          repoUrl: `https://${json['repository']['url'].match('.*(github\.com.*)\.git')[1]}`,
-          currentVersionDate: json['time'][currentVersion],
-          targetVersionDate: json['time'][targetVersion],
+          repoUrl: `https://${
+            json["repository"]["url"].match(".*(github.com.*).git")[1]
+          }`,
+          currentVersionDate: json["time"][currentVersion],
+          targetVersionDate: json["time"][targetVersion],
         });
-      }
-      else {
+      } else {
         reject(`Repository not found for ${homepageUrl}`);
       }
     });
   });
 }
 
-
 // retrieve data
 // ===================
 
-function checkNextDependency(dependencies) {
-  let allKeys = Object.keys(dependencies);
-
-  if (allKeys.length === 0) {
-    writeFiles();
-    return;
-  }
-
-  let key = allKeys[0];
-  let currentVersion = dependencies[key]['current'];
-  let targetVersion = dependencies[key]['latest'];
-  let homepageUrl = dependencies[key]['homepage'];
+async function checkDependency(i, key, dependency) {
+  const currentVersion = dependency["current"] || dependency["wanted"];
+  const targetVersion = dependency["latest"];
+  const homepageUrl = dependency["homepage"];
   let registryInfos;
 
-  console.log(``);
-  logMessage(`${key} (${changelogs.length + 1}/${allKeys.length + changelogs.length})`);
+  try {
+    registryInfos = await getRegistryInfos(
+      homepageUrl,
+      key,
+      currentVersion,
+      targetVersion
+    );
+    const repoRawUrl = await getGithubRawUrl(registryInfos.repoUrl);
+    const changelog = await loadChangelog(
+      registryInfos,
+      repoRawUrl,
+      currentVersion
+    );
 
-  getRegistryInfos(homepageUrl, key, currentVersion, targetVersion)
-    .then((infos) => {
-      registryInfos = infos;
-      return getGithubRawUrl(registryInfos.repoUrl);
-    })
-    .then((repoRawUrl) => {
-      return loadChangelog(registryInfos, repoRawUrl, currentVersion)
-    })
-    .then((changelog) => {
-      let result = formatChangelog(key, changelog, currentVersion, targetVersion, registryInfos.repoUrl);
-      changelogs.push(result);
-      return sleep(1000);
-    })
-    .then(() => {
-      logSuccess(`✔ ${key} done`);
-      delete dependencies[key];
-      checkNextDependency(dependencies);
-    })
-    .catch((error) => {
-      logError(`fail for ${key} with ${error}`);
-      const result = formatChangelog(`${key}`, 'Changelog not found', currentVersion, targetVersion, registryInfos ? registryInfos.repoUrl : undefined);
-      changelogs.push(result);
-      delete dependencies[key];
-      checkNextDependency(dependencies);
-    })
+    const result = formatChangelog(
+      i,
+      key,
+      changelog,
+      currentVersion,
+      targetVersion,
+      registryInfos.repoUrl
+    );
+
+    logSuccess(`✔ ${key} done`);
+    return result;
+  } catch (error) {
+    const result = formatChangelog(
+      i,
+      `${key}`,
+      "Changelog not found",
+      currentVersion,
+      targetVersion,
+      registryInfos ? registryInfos.repoUrl : undefined
+    );
+
+    logError(`fail for ${key} with ${error}`);
+    return result;
+  }
 }
 
-async function loadChangelog(registryInfos, githubUrl, version, pos=0) {
-  let endpoints = ['CHANGELOG.md', 'RELEASENOTES.md'];
-  let url = `${githubUrl}${endpoints[pos]}`;
+async function loadChangelog(registryInfos, githubUrl, version, pos = 0) {
+  const endpoints = ["CHANGELOG.md", "RELEASENOTES.md"];
+  const url = `${githubUrl}${endpoints[pos]}`;
   logInfo(`get changelog at ${url} (v${version})`);
 
-	return new Promise((resolve, reject) => {
+  return new Promise((resolve, reject) => {
     request(url, {}, (err, res, body) => {
       if (res && res.statusCode < 300) {
         resolve(body);
-      }
-      else {
-        if (pos >= endpoints.length-1) {
+      } else {
+        if (pos >= endpoints.length - 1) {
           reject(`Changelog not found for ${githubUrl}`);
         } else {
-          loadChangelog(registryInfos, githubUrl, version, pos+1).then((content) => {
-            resolve(content);
-          }).catch((error) => {
-            getCommits(registryInfos).then((content) => {
+          loadChangelog(registryInfos, githubUrl, version, pos + 1)
+            .then((content) => {
               resolve(content);
-            }).catch((error) => {
-              reject(error);
+            })
+            .catch((error) => {
+              getCommits(registryInfos)
+                .then((content) => {
+                  resolve(content);
+                })
+                .catch((error) => {
+                  reject(error);
+                });
             });
-          });
         }
-      }
-    })
-  });
-}
-
-function getGithubRawUrl(url) {
-	return new Promise((resolve, reject) => {
-    if(isGithubUrl(url)) {
-      let rawUrl = url;
-      let anchorPos = url.indexOf('#');
-      if (anchorPos !== -1) rawUrl = url.slice(0, anchorPos);
-      rawUrl = rawUrl.replace('https://github.com', 'https://raw.githubusercontent.com');
-      resolve(`${rawUrl}/master/`);
-    } else {
-      reject('not a github repo');
-    }
-  });
-}
-
-function isGithubUrl(url) {
-  return url.match('https://github.com') !== null;
-}
-
-function getRegistryUrl(dependencyName) {
-  let url = lockFile['dependencies'][dependencyName]['resolved'] || '';
-  let dashPos = url.indexOf('/-/');
-  return url.slice(0, dashPos);
-}
-
-async function getCommits(registryInfos) {
-	return new Promise((resolve, reject) => {
-    let repoName = registryInfos.repoUrl.match('.*github\.com\/(.*)')[1];
-    logInfo(`get commits for ${repoName}`);
-
-    let url = `https://api.github.com/repos/${repoName}/commits?since=${registryInfos.currentVersionDate}&until=${registryInfos.targetVersionDate}`;
-    request(url, { json: true, headers: {'User-Agent': 'odc-1.0.0'} }, (err, res, json) => {
-      if (res && res.statusCode < 300 && json) {
-        resolve(json.map((res) => `• ${res.commit.committer.date} - ${res.commit.message.replace(/\\n/g, ' | ')}`).join('\n'));
-      }
-      else {
-        reject(`Commit not found for ${repoName}`);
       }
     });
   });
 }
 
+async function getGithubRawUrl(url) {
+  return new Promise((resolve, reject) => {
+    if (isGithubUrl(url)) {
+      let rawUrl = url;
+      const anchorPos = url.indexOf("#");
+      if (anchorPos !== -1) rawUrl = url.slice(0, anchorPos);
+      rawUrl = rawUrl.replace(
+        "https://github.com",
+        "https://raw.githubusercontent.com"
+      );
+      resolve(`${rawUrl}/master/`);
+    } else {
+      reject("not a github repo");
+    }
+  });
+}
+
+function isGithubUrl(url) {
+  return url.match("https://github.com") !== null;
+}
+
+function getRegistryUrl(dependencyName) {
+  const url = lockFile["dependencies"][dependencyName]["resolved"] || "";
+  const dashPos = url.indexOf("/-/");
+  return url.slice(0, dashPos);
+}
+
+async function getCommits(registryInfos) {
+  return new Promise((resolve, reject) => {
+    const repoName = registryInfos.repoUrl.match(".*github.com/(.*)")[1];
+    logInfo(`get commits for ${repoName}`);
+
+    const url = `https://api.github.com/repos/${repoName}/commits?since=${registryInfos.currentVersionDate}&until=${registryInfos.targetVersionDate}`;
+    request(
+      url,
+      { json: true, headers: { "User-Agent": "odc-1.0.1" } },
+      (err, res, json) => {
+        if (res && res.statusCode < 300 && json) {
+          resolve(
+            json
+              .map(
+                (res) =>
+                  `• ${
+                    res.commit.committer.date
+                  } - ${res.commit.message.replace(/\\n/g, " | ")}`
+              )
+              .join("\n")
+          );
+        } else {
+          reject(`Commit not found for ${repoName}`);
+        }
+      }
+    );
+  });
+}
 
 // template
 // ===================
 
 function executeTemplate(filePath, data = {}) {
-  const template = fs.readFileSync(filePath, 'utf8');
+  const template = fs.readFileSync(filePath, "utf8");
   const regex = /{{\s*(.+?)\s*}}/g;
 
   let output = template;
-  while (match = regex.exec(template)) {
+  while ((match = regex.exec(template))) {
     output = output.replace(match[0], eval(`data.${match[1]}`) || "");
   }
 
   return output;
 }
 
-function formatChangelog(dependencyName, content, currentVersion, targetVersion, url) {
-  let html = converter.makeHtml(content);
-  html = html.replace(/<pre><code/g, '<pre class="code"><code');
+function formatChangelog(
+  id,
+  dependencyName,
+  content,
+  currentVersion,
+  targetVersion,
+  url
+) {
+  const html = converter
+    .makeHtml(content)
+    .replace(/<pre><code/g, '<pre class="code"><code');
 
   return {
-    id: changelogs.length + 1,
-    url: url,
+    id,
+    url,
     name: dependencyName,
-    filename: `${dependencyName}_${currentVersion}-${targetVersion}.html`.replace(/\//g, '-'),
+    filename: `${dependencyName}_${currentVersion}-${targetVersion}.html`.replace(
+      /\//g,
+      "-"
+    ),
     content: html,
     version: {
       current: currentVersion,
       target: targetVersion,
       type: getUpgradeType(currentVersion, targetVersion),
     },
-  }
+  };
 }
 
 function getUpgradeType(currentVersion = '', targetVersion = '') {
@@ -251,36 +306,42 @@ function getUpgradeType(currentVersion = '', targetVersion = '') {
   }
 }
 
-function writeFiles() {
+function writeFiles(allChangelogList) {
   console.log(``);
   logMessage(`=====================`);
 
-  fs.copyFileSync(`${staticDirectory}/index.css`, `${outputDirectory}/index.css`);
+  fs.copyFileSync(
+    `${staticDirectory}/index.css`,
+    `${outputDirectory}/index.css`
+  );
 
-  for (const changelogData of changelogs) {
+  for (const changelogData of allChangelogList) {
     logMessage(`write ${changelogData.filename}`);
 
-    let nav = '';
-    for (const navData of changelogs) {
+    let nav = "";
+    for (const navData of allChangelogList) {
       nav += executeTemplate(`${staticDirectory}/nav-template.html`, {
-        changelog: navData, 
-        selectedClass: navData.name === changelogData.name ? 'selected' : ''
+        changelog: navData,
+        selectedClass: navData.name === changelogData.name ? "selected" : "",
       });
     }
 
-    const html = executeTemplate(`${staticDirectory}/main-template.html`, { 
-      nav: nav, changelog: 
-      changelogData
+    const html = executeTemplate(`${staticDirectory}/main-template.html`, {
+      nav: nav,
+      changelog: changelogData,
     });
 
     fs.writeFileSync(`${outputDirectory}/${changelogData.filename}`, html);
   }
 
-  console.log("")
-  logSuccess(`${changelogs.length} changelog${changelogs.length > 1 ? 's' : ''} available in ${outputDirectory}`);
+  console.log("");
+  logSuccess(
+    `${allChangelogList.length} changelog${
+      allChangelogList.length > 1 ? "s" : ""
+    } available in ${outputDirectory}`
+  );
   logSuccess(`Ready ! 🚀`);
 }
-
 
 // Helpers
 // ===================
@@ -302,11 +363,11 @@ function logError(message) {
 }
 
 function sleep(ms) {
-  return new Promise(resolve => setTimeout(resolve, ms));
+  return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
 function clearOutputDirectory() {
   logMessage(`clear output directory ${outputDirectory}`);
-	rimraf.sync(outputDirectory);
+  rimraf.sync(outputDirectory);
   fs.mkdirSync(outputDirectory, { recursive: true });
 }